backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
chromadb
langchain-core
langchain-community
langchain-openai
pydantic
